pull music links out of comments on all threads
Logic for the rake task previously did not take into account
comments on posts that did not include a video. this terrible
pile of code that I whipped up in two hours attempts to correct
that. This is bad, so very bad, but I have a date to get to right
now and it needs to work.